// report-ng/app-react/src/model/report-model/framework_pb.ts
export const ClickPathEventType = {
  CPET_NOT_SET: 0,
  CPET_WINDOW: 1,
  CPET_CLICK: 2,
  CPET_VALUE: 3,
  CPET_PAGE: 4,
  CPET_URL: 5,
} as const;

export type ClickPathEventType = typeof ClickPathEventType[keyof typeof ClickPathEventType];
// ...
export const ResultStatusType = {
  RST_NOT_SET: 0,
  /** NO_RUN - default status when entity is spawned, basically an illegal end status */
  NO_RUN: 1,
  /**
   * INFO - info status, not representative
   *
   * @deprecated
   */
  INFO: 2,
  /** SKIPPED - skipped */
  SKIPPED: 3,
  /** PASSED - passed status, without any issue */
  PASSED: 4,
  /**
   * MINOR - passed, but with minor, non-blocking issues
   *
   * @deprecated
   */
  MINOR: 5,
  /** FAILED - failed status, with hard failing issues */
  FAILED: 7,
  /**
   * FAILED_MINOR - failed with additional minor issues
   *
   * @deprecated
   */
  FAILED_MINOR: 8,
  /** FAILED_RETRIED - failed and retry is triggered */
  FAILED_RETRIED: 9,
  /** FAILED_EXPECTED - entity was expected to fail (with or without an explicit issue), it may or may not be representative */
  FAILED_EXPECTED: 10,
  /** PASSED_RETRY - passed after a retry */
  PASSED_RETRY: 11,
  /**
   * MINOR_RETRY - minor after a retry
   *
   * @deprecated
   */
  MINOR_RETRY: 12,
  /** REPAIRED - passed with fail annotation */
  REPAIRED: 13,
} as const;

export type ResultStatusType = typeof ResultStatusType[keyof typeof ResultStatusType];
// ...
  export type NO_RUN = typeof ResultStatusType.NO_RUN;
// ...
export interface TestStep {
  name?:
    | string
    | undefined;
  /** string id = 2; */
  actions?: TestStepAction[] | undefined;
}

export interface TestStepAction {
  name?:
    | string
    | undefined;
  /** string id = 2; */
  timestamp?:
    | number
    | undefined;
  /**
   * repeated string screenshot_names = 4 [deprecated = true];
   *    repeated ClickPathEvent clickpath_events = 5 [deprecated = true];
   *    repeated string screenshot_ids = 6 [deprecated = true];
   */
  entries?: TestStepActionEntry[] | undefined;
}

export interface TestStepActionEntry {
  clickPathEvent?: ClickPathEvent | undefined;
  screenshotId?:
    | string
    | undefined;
  /** @deprecated */
  logMessage?:
    | LogMessage
    | undefined;
  /** @deprecated */
  assertion?: ErrorContext | undefined;
  errorContext?: ErrorContext | undefined;
  logMessageId?: string | undefined;
}

export interface ClickPathEvent {
  type?: ClickPathEventType | undefined;
  subject?: string | undefined;
  sessionId?: string | undefined;
}

export interface LogMessage {
  type?: LogMessageType | undefined;
  loggerName?: string | undefined;
  message?: string | undefined;
  timestamp?: number | undefined;
  threadName?: string | undefined;
  stackTrace?: StackTraceCause[] | undefined;
  prompt?: boolean | undefined;
  id?: string | undefined;
}

export interface ErrorContext {
  /**
   * string readable_error_message = 1 [deprecated = true];
   *    string additional_error_message = 2 [deprecated = true];
   *    StackTrace stack_trace = 3  [deprecated = true];
   *    string error_fingerprint = 6 [deprecated = true];
   */
  scriptSource?:
    | ScriptSource
    | undefined;
  /**
   * ScriptSource execution_object_source = 8 [deprecated = true];
   *
   * @deprecated
   */
  ticketId?:
    | string
    | undefined;
  /** @deprecated */
  description?: string | undefined;
  stackTrace?: StackTraceCause[] | undefined;
  optional?: boolean | undefined;
  id?: string | undefined;
}
// ...
export const TestStep: MessageFns<TestStep> = {
  encode(message: TestStep, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.name !== undefined && message.name !== "") {
      writer.uint32(10).string(message.name);
    }
    if (message.actions !== undefined && message.actions.length !== 0) {
      for (const v of message.actions) {
        TestStepAction.encode(v!, writer.uint32(26).fork()).join();
      }
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): TestStep {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseTestStep();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.name = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          const el = TestStepAction.decode(reader, reader.uint32());
          if (el !== undefined) {
            message.actions!.push(el);
          }
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};
// ...
export const ClickPathEvent: MessageFns<ClickPathEvent> = {
  encode(message: ClickPathEvent, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.type !== undefined && message.type !== 0) {
      writer.uint32(8).int32(message.type);
    }
    if (message.subject !== undefined && message.subject !== "") {
      writer.uint32(18).string(message.subject);
    }
    if (message.sessionId !== undefined && message.sessionId !== "") {
      writer.uint32(26).string(message.sessionId);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ClickPathEvent {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseClickPathEvent();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.type = reader.int32() as any;
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.subject = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.sessionId = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};

function createBaseLogMessage(): LogMessage {
  return { type: 0, loggerName: "", message: "", timestamp: 0, threadName: "", stackTrace: [], prompt: false, id: "" };
}

export const LogMessage: MessageFns<LogMessage> = {
  encode(message: LogMessage, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.type !== undefined && message.type !== 0) {
      writer.uint32(8).int32(message.type);
    }
    if (message.loggerName !== undefined && message.loggerName !== "") {
      writer.uint32(18).string(message.loggerName);
    }
    if (message.message !== undefined && message.message !== "") {
      writer.uint32(26).string(message.message);
    }
    if (message.timestamp !== undefined && message.timestamp !== 0) {
      writer.uint32(32).int64(message.timestamp);
    }
    if (message.threadName !== undefined && message.threadName !== "") {
      writer.uint32(42).string(message.threadName);
    }
    if (message.stackTrace !== undefined && message.stackTrace.length !== 0) {
      for (const v of message.stackTrace) {
        StackTraceCause.encode(v!, writer.uint32(50).fork()).join();
      }
    }
    if (message.prompt !== undefined && message.prompt !== false) {
      writer.uint32(56).bool(message.prompt);
    }
    if (message.id !== undefined && message.id !== "") {
      writer.uint32(66).string(message.id);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): LogMessage {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseLogMessage();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.type = reader.int32() as any;
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.loggerName = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.message = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.timestamp = longToNumber(reader.int64());
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.threadName = reader.string();
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          const el = StackTraceCause.decode(reader, reader.uint32());
          if (el !== undefined) {
            message.stackTrace!.push(el);
          }
          continue;
        }
        case 7: {
          if (tag !== 56) {
            break;
          }

          message.prompt = reader.bool();
          continue;
        }
        case 8: {
          if (tag !== 66) {
            break;
          }

          message.id = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};

function createBaseErrorContext(): ErrorContext {
  return { scriptSource: undefined, ticketId: "", description: "", stackTrace: [], optional: false, id: "" };
}

export const ErrorContext: MessageFns<ErrorContext> = {
  encode(message: ErrorContext, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.scriptSource !== undefined) {
      ScriptSource.encode(message.scriptSource, writer.uint32(58).fork()).join();
    }
    if (message.ticketId !== undefined && message.ticketId !== "") {
      writer.uint32(74).string(message.ticketId);
    }
    if (message.description !== undefined && message.description !== "") {
      writer.uint32(82).string(message.description);
    }
    if (message.stackTrace !== undefined && message.stackTrace.length !== 0) {
      for (const v of message.stackTrace) {
        StackTraceCause.encode(v!, writer.uint32(90).fork()).join();
      }
    }
    if (message.optional !== undefined && message.optional !== false) {
      writer.uint32(96).bool(message.optional);
    }
    if (message.id !== undefined && message.id !== "") {
      writer.uint32(106).string(message.id);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ErrorContext {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseErrorContext();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 7: {
          if (tag !== 58) {
            break;
          }

          message.scriptSource = ScriptSource.decode(reader, reader.uint32());
          continue;
        }
        case 9: {
          if (tag !== 74) {
            break;
          }

          message.ticketId = reader.string();
          continue;
        }
        case 10: {
          if (tag !== 82) {
            break;
          }

          message.description = reader.string();
          continue;
        }
        case 11: {
          if (tag !== 90) {
            break;
          }

          const el = StackTraceCause.decode(reader, reader.uint32());
          if (el !== undefined) {
            message.stackTrace!.push(el);
          }
          continue;
        }
        case 12: {
          if (tag !== 96) {
            break;
          }

          message.optional = reader.bool();
          continue;
        }
        case 13: {
          if (tag !== 106) {
            break;
          }

          message.id = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};